+++ client/src/types/common.ts
@@ -1,14 +1,8 @@
-import type { Color, FontSize, BreakPoint } from '~styles/theme';
-
-export { Color, FontSize, BreakPoint };
+import type { Color } from '~styles/theme';
 
 export type GlobalColors = keyof Color;
 
 export type GlobalSizes = 'x-small' | 'small' | 'medium' | 'large' | 'x-large';
-
-export type CommonBasicValues = 'inherit' | 'initial' | 'unset';
-
-export type PercentNumber = `${number}%`;
 
 export interface ImageObject {
   src: string;
